refactoring - moved contract models into Contracts folder
