+++ ShamelessShiftGrabber/Macrodroid/IncomingShiftExtensions.cs
@@ -1,3 +1,5 @@
+using ShamelessShiftGrabber.Contracts;
+
 namespace ShamelessShiftGrabber.Macrodroid;
 
 internal static class IncomingShiftExtensions
